Refactor AI provider wiring
